+++ core/components/romanesco/docs/changelog.md
@@ -8,6 +8,7 @@ New features:
 - Add lazy loading to Image CB and image overviews
 
 Fixes and improvements:
+- Add suffix with version number to main CSS and JS assets
 - Optimize caching of chunks in Overview templates
 - Generate responsive content images with srcset and sizes
 - Add round and removeDuplicateLines snippets